Migrate prototype API to v1 API
