--- src/noteburst/handlers/prototyping/__init__.py
@@ -1,3 +1,0 @@
-from .handlers import prototype_router
-
-__all__ = ["prototype_router"]
